fixed type for oemval2str (shouldn't be 16 bits for a 24 bits iana ..)

+++ ipmitool/include/ipmitool/helper.h
@@ -63,7 +63,7 @@ struct oemvalstr {
 };
 
 const char * val2str(uint16_t val, const struct valstr * vs);
-const char * oemval2str(uint16_t oem,uint16_t val, const struct oemvalstr * vs);
+const char * oemval2str(uint32_t oem,uint16_t val, const struct oemvalstr * vs);
 uint16_t str2val(const char * str, const struct valstr * vs);
 void print_valstr(const struct valstr * vs, const char * title, int loglevel);
 void print_valstr_2col(const struct valstr * vs, const char * title, int loglevel);
